Create telegram_custom_alert.py
This script takes a number of 1 minute candles, defined by the user, and retrieves the biggest closing price. If the current price of the coin is bigger than the biggest closing price, an alert is sent to your telegram chat with the bot.

In order to find your chat id, after you give the bot the /start command in telegram, send him another message and head to this link (also delete <  > from the link):
https://api.telegram.org/bot<YOUR API TOKEN>/getUpdates

On this page you will see a message similar to this:
"message":{"message_id":3,"from":{"id":1234355312,"is_bot":false,"first_name":"test","language_code":"en"},"chat":{"id":555555555,"first_name":"test","type":"private"},"date":1617047587,"text":"hello"}}]}
As you can see the chat id in this case is 555555555. Just copy yours and paste it in the code